Json 불러오는 기능 추가 드디어 좀만 더 하면 애니메이션도

+++ Windows_Programming_Final_Project/GameEngine/FileBase/FileTypes/Sprite.cpp
@@ -3,7 +3,7 @@
 #include "Sprite.h"
 #include "Texture.h"
 
-Sprite::Sprite(Texture* texture, int x, int y, int cx, int cy)
+Sprite::Sprite(Texture* texture, float x, float y, float cx, float cy)
 	: _texture(texture), _x(x), _y(y), _cx(cx), _cy(cy)
 {
